+++ test/bigtest/tests/instance-movement-page-test.js
@@ -16,7 +16,7 @@ const marcRecord = {
   },
 };
 
-describe('InstanceMovementPage', () => {
+describe.only('InstanceMovementPage', () => {
   setupApplication();
 
   const instanceMarcPage = new InstanceMarcPage();
@@ -76,6 +76,7 @@ describe('InstanceMovementPage', () => {
 
   it('should render instance details to', () => {
     expect(instanceMovementDetailsTo.isPresent).to.be.true;
+    expect(instanceMovementDetailsTo.emptyDropZone.isPresent).to.be.true;
     expect(instanceMovementDetailsTo.title.includes(titleInstanceTo)).to.be.true;
   });